feat: delete ProjectIconBox as redundant, implement IconWithTooltip directly in ProjectHeader component; update /components/index.ts exports
