+++ src/components/ProjectHeader.tsx
@@ -2,7 +2,8 @@ import  CardHeader from "@mui/material/CardHeader";
 
 import { Project } from "../types";
 import { icons } from "../data/icons";
-import { IconWithTooltip, ProjectIconBox } from "./index";
+import { IconWithTooltip } from "./index";
+
 
 
 const unfoldIconOptions = {
@@ -16,6 +17,29 @@ const unfoldIconOptions = {
     },
   };
 
+  const githubIconOptions = (githubLink: string) => ({
+    icon: {
+      ariaLabel: 'View on GitHub',
+      icon: icons.GithubBlack,
+    }, 
+    link: {
+      ariaLabel: 'View on GitHub',
+      url: githubLink,
+    }
+  });
+  
+  const liveLinkOptions = (liveLink: string) => ({
+    icon: { 
+      ariaLabel: 'Navigate to live website',
+      icon: icons.ExternalLink,
+    },
+    link: {
+      ariaLabel: 'Navigate to live website',
+      url: liveLink ?? undefined,
+    }
+  });
+  
+
 type Props = {
     project: Project,
     fullDisplay: boolean,
@@ -26,9 +50,6 @@ type Props = {
 
 export const ProjectHeader: React.FC<Props> = ({ project, fullDisplay, toggleDisplay, index }) => {
     const { title, githubLink, liveLink, subheader } = project;
-  
-    const projectLinks = { liveLink, githubLink };
-
 
     const handleToggleDisplay = () => {
         toggleDisplay(index);
@@ -40,7 +61,6 @@ export const ProjectHeader: React.FC<Props> = ({ project, fullDisplay, toggleDis
           <span> 
             <IconWithTooltip
               icon={unfoldIconOptions[fullDisplay.toString()]}
-              onClick={handleToggleDisplay}
               placement='right'
             />
           </span>
@@ -49,18 +69,24 @@ export const ProjectHeader: React.FC<Props> = ({ project, fullDisplay, toggleDis
 
 
     return (
-        <CardHeader
-        className='bg-whitesmoke border-radius-4px'
-         action={
-           <ProjectIconBox
-             projectLinks={projectLinks}
-             fullDisplay={fullDisplay}
-             toggleDisplay={toggleDisplay}
-             index={index}
-           />
+       <div onClick={handleToggleDisplay}>
+         <CardHeader
+          className='bg-whitesmoke border-radius-4px'
+          action={
+            <div className="display-flex flex-center center-column">
+              <IconWithTooltip {...githubIconOptions(githubLink)} placement='right'/>
+             {
+              liveLink && 
+                <IconWithTooltip
+                  {...liveLinkOptions(liveLink)} 
+                  placement='right'
+                />
+              }
+            </div>
          }
-         title={titleBox}
-         subheader={subheader}
+          title={titleBox}
+          subheader={subheader}
        />
+       </div>
     )
 }
