+++ src/components/index.ts
@@ -3,8 +3,7 @@ export * from './DisplayFull';
 export * from './DisplayPreview/DisplayPreview';
 export * from './Footer';
 export * from './Form';
-export * from './IconWithTooltip/IconWithTooltip';
-export * from './IconWithTooltip/ProjectIconBox';
+export * from './IconWithTooltip';
 export * from './ImageContainer';
 export * from './Loading/Loading';
 export * from './SideBySideSection';
